zhang
调整删除发展历程节点

+++ manage/develop/del.php
@@ -7,7 +7,9 @@ include('../../inc/manage/config.php');
 
 $DevID = htmlentities($_GET['DevID']);
 $info = $db->get_one('develop', "DevID=$DevID");
-unlink($site_root_path.$info['pic_src']);
+if (!empty($info['pic_src'])) {
+	unlink($site_root_path.$info['pic_src']);
+}
 $db->delete('develop', "DevID=$DevID");
 
 header('Location:index.php');
